### 2019-11-28 #1921 @whedon
[PRE REVIEW]: Pyglmnet: Python implementation of elastic-net regularized generalized linear models
**Submitting author:** @jasmainak (<a href="http://orcid.org/0000-0002-3199-9027">Mainak Jas</a>)
**Repository:** <a href="https://github.com/glm-tools/pyglmnet/" target ="_blank">https://github.com/glm-tools/pyglmnet/</a>
**Version:** v1.1
**Reviewers:** professoralkmin, ryEllison
