### 2019-12-02 d98fd9a
Add missing DOIs to paper.
