+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={5},
   pages={e0196527},
   year={2018},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0196527}
 }
 
 @article{rybakken2019decoding,
@@ -17,7 +18,8 @@
   number={1},
   pages={68--93},
   year={2019},
-  publisher={MIT Press}
+  publisher={MIT Press},
+  doi={10.1162/neco_a_01150}
 }
 
 
@@ -27,7 +29,8 @@
   journal={bioRxiv},
   pages={609826},
   year={2019},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/609826}
 }
 
 @article{benjamin2017modern,
@@ -36,7 +39,8 @@
   journal={bioRxiv},
   pages={111450},
   year={2017},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/111450}
 }
 
 @inproceedings{sklearn_api,
