### 2019-12-02 #1921 @jasmainak
Thanks to @geektoni, the missing DOIs should be fixed now!
