### 2019-12-02 #1921 @arokem
👋 ryEllison : would you be willing to review this paper for JOSS?
