### 2019-12-02 #1921 @arokem
👋 @b-freitas : would you be willing to review this article for JOSS?
