### 2019-12-12 #1921 @arokem
:wave: @nmoran : would you be willing to review this  paper for JOSS?
