### 2019-12-12 #1921 @arokem
:wave: @zoomelectrico: would you be willing to review this paper for JOSS?
