### 2019-12-16 #1921 @arokem
👋 @prateekiiest: would you be willing to review this paper for JOSS?
