### 2019-12-16 #1921 @arokem
👋 @professoralkmin: would you be willing to review this paper for JOSS?
