### 2019-12-16 #1921 @ryEllison
Yes.

R

Sent from my iPhone

> On Dec 2, 2019, at 2:35 PM, Ariel Rokem <notifications@github.com> wrote:
>
> ﻿
> 👋 @ryEllison : would you be willing to review this article for JOSS?
>
