### 2019-12-16 #1921 @professoralkmin
Yes

Em seg, 16 de dez de 2019 10:31 AM, ryEllison <notifications@github.com>
escreveu:

> Yes.
>
> R
>
> Sent from my iPhone
>
> > On Dec 2, 2019, at 2:35 PM, Ariel Rokem <notifications@github.com>
> wrote:
> >
> > ﻿
> > 👋 @ryEllison : would you be willing to review this article for JOSS?
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub, or unsubscribe.
>
