### 2019-12-16 #1959 @whedon
[REVIEW]: Pyglmnet: Python implementation of elastic-net regularized generalized linear models
**Submitting author:** @jasmainak (<a href="http://orcid.org/0000-0002-3199-9027">Mainak Jas</a>)
**Repository:** <a href="https://github.com/glm-tools/pyglmnet/" target ="_blank">https://github.com/glm-tools/pyglmnet/</a>
**Version:** 1.1.1
