### 2020-01-10 #1959 @professoralkmin
Hello @arokem. Yes, I read and tested everything. Now, I already finished and ticked all according to my work experience and academic technology life. Thanks. :)
