### 2020-01-14 #1959 @ryEllison
I'm happy to see this paper published as is. I believe it's well written with adequate documentation and clearly expands on incomplete GLM packages in the machine learning domain.

I would like to point out that the example code in your README file does GLM estimation but doesn't give the person trying to install and use your package any sort of result. Nothing is printed to the terminal. No visualizations are created. Although, the objects holding this information are there and someone could certainly run a script with this code in it interactively using the `-i` flag and 'back-of-the-envelope' spit out some results - or I suppose, more prudently (haha), alter the code to output what you'd like to see - this shouldn't be the responsibility of the end-user, in my opinion, when trying to look at your documentation.

Otherwise, great work! Congrats!
