### 2020-01-20 #1959 @arokem
Sorry about the slowness here: @professoralkmin : I see that the several boxes are still not ticked in your section. Do you still have concerns with these requirements?
