### 2020-01-21 #1959 @professoralkmin
@arokem I believe that it is necessary 1) to explain better about the example usage to be more clear how to use correctly the application and 2) to review of some sentences to be improve the quality of writing.
