### 2020-01-24 de9b4c6
Add small example of pyglmnet usage in the paper.
