+++ paper/paper.bib
@@ -55,3 +55,10 @@
   year      = {2013},
   pages = {108--122},
 }
+
+@misc{Dua:2019,
+author = "Dua, Dheeru and Graff, Casey",
+year = "2017",
+title = "{UCI} Machine Learning Repository",
+url = "http://archive.ics.uci.edu/ml",
+institution = "University of California, Irvine, School of Information and Computer Sciences" }
