+++ paper/paper.md
@@ -172,6 +172,28 @@ As a result of this compatibility, ``scikit-learn`` tools for building pipelines
 [@bertran2018active; @rybakken2019decoding; @hofling2019probing; @benjamin2017modern]. It contains unit tests and includes documentation in the form of tutorials, docstrings and
 examples that are run through continuous integration.
 
+# Example Usage
+We present now how to use pyglmnet on a real-world problem. We show how to apply our
+library to the Community and Crime dataset [@Dua:2019]. We predict the per capita violent
+crimes by employing demographic data comprising 128 attributes and by using a Binomial
+Distributed GLM with Elastic-Net regularization.
+
+```py
+from sklearn.model_selection import train_test_split
+from pyglmnet import GLMCV, simulate_glm, datasets
+
+# Read dataset and split it into train/test
+X, y = datasets.fetch_community_crime_data()
+Xtrain, Xtest, ytrain, ytest = train_test_split(X, y, test_size=0.33)
+
+# Generate a binomial distributed GLM with elastic-net regularization
+glm = GLMCV(distr='binomial', alpha=0.05, score_metric='pseudo_R2', cv=3)
+
+# Fit the model and then predict
+glm.fit(Xtrain, ytrain)
+yhat = glm.predict_proba(Xtest)
+```
+
 # Acknowledgements
 
 ``Pyglmnet`` development is partly supported by NIH NINDS R01-NS104585 and the Special Research Fund (B.O.F.) of Ghent University.
