### 2020-02-03 #1959 @arokem
Hi @jasmainak: have you had a chance to address these comments? In particular, I tend to agree with @professoralkmin's comment that more detail is needed in the example usage. These include brief introductions that provide context, but there is not enough detail to understand the various steps taken and to understand the outputs. I'd suggest adding these details.
