### 2020-02-09 7723f8f
MAINT: try to freeze numpy
