### 2020-02-09 e2bf1de
MAINT use int for linspace
