+++ examples/plot_tikhonov.py
@@ -75,7 +75,7 @@ strf_model.visualize_gaussian_basis(spatial_basis)
 ########################################################
 # Design temporal basis functions
 
-time_points = np.linspace(-100., 100., 10.)
+time_points = np.linspace(-100., 100., 10)
 centers = [-75., -50., -25., 0, 25., 50., 75.]
 temporal_basis = strf_model.make_raised_cosine_temporal_basis(
     time_points=time_points,
