### 2020-02-09 f2e24f0
more focused description of differentiators; wordsmithing examples
