### 2020-02-09 6ffd635
more refactoring; link to documentation
