+++ paper/paper.bib
@@ -58,7 +58,7 @@
 
 @misc{Dua:2019,
 author = "Dua, Dheeru and Graff, Casey",
-year = "2017",
+year = "2019",
 title = "{UCI} Machine Learning Repository",
 url = "http://archive.ics.uci.edu/ml",
 institution = "University of California, Irvine, School of Information and Computer Sciences" }
