### 2020-02-17 13aa027
Update paper.md, mailmap and small URL fix
