### 2020-02-17 3c824f5
Update paper.md, mailmap and small URL fix
