### 2020-02-17 c935175
Update README and paper text
