+++ doc/README.md
@@ -2,9 +2,9 @@
 
 Please see full documentation page [here](http://glm-tools.github.io/pyglmnet/).
 
-We use `sphinx` to generate documentation page.
+We use [`sphinx`](https://www.sphinx-doc.org/en/master/) with [`sphinx-rtd-theme`](https://sphinx-rtd-theme.readthedocs.io/en/stable/) to generate a documentation page.
 
 To build documentation page, run `make html`. All static files will be built in
 `_build/html/` where you can open them using the web browser.
 
-To push built documentation page to `gh-pages`, simply run `make install`
+To push built documentation page to `gh-pages` branch, simply run `make install`
