### 2020-02-17 d804ddc
Update README and paper text
