+++ paper/paper.bib
@@ -57,8 +57,9 @@
 }
 
 @misc{Dua:2019,
-author = "Dua, Dheeru and Graff, Casey",
-year = "2019",
-title = "{UCI} Machine Learning Repository",
-url = "http://archive.ics.uci.edu/ml",
-institution = "University of California, Irvine, School of Information and Computer Sciences" }
+  author = "Dua, Dheeru and Graff, Casey",
+  year = "2019",
+  title = "{UCI} Machine Learning Repository",
+  url = "http://archive.ics.uci.edu/ml",
+  institution = "University of California, Irvine, School of Information and Computer Sciences"
+}
