### 2020-02-17 9385101
Further update sentences in paper.md
