### 2020-02-17 e09b9a1
minor wordsmithing changes
