+++ doc/README.md
@@ -1,10 +1,10 @@
 # pyglmnet documentation
 
-Please see full documentation page [here](http://glm-tools.github.io/pyglmnet/).
+Please see the full documentation page [here](http://glm-tools.github.io/pyglmnet/).
 
 We use [`sphinx`](https://www.sphinx-doc.org/en/master/) with [`sphinx-rtd-theme`](https://sphinx-rtd-theme.readthedocs.io/en/stable/) to generate a documentation page.
 
-To build documentation page, run `make html`. All static files will be built in
+To build the documentation page, run `make html`. All static files will be built in
 `_build/html/` where you can open them using the web browser.
 
-To push built documentation page to `gh-pages` branch, simply run `make install`
+To push the built documentation page to the `gh-pages` branch, simply run `make install`
