### 2020-02-17 17cfb07
Minor fix in paper.md
