### 2020-02-18 46c0c23
DOC: improve readme example
