### 2020-02-18 f4f8ad4
DOC: need matplotlib for intro example now
