+++ .travis.yml
@@ -16,6 +16,7 @@ install:
     - source activate testenv
     - conda install --yes --file requirements.txt
     - pip install --upgrade twine
+    - pip install matplotlib==3.0
     - python setup.py develop
     - python setup.py sdist bdist_wheel
 
