### 2020-02-18 3143755
FIX: don't use use_line_collection as it's not in matplotlib 3.0 and matplotlib 3.1 is not compatible with python 3.5
