### 2020-02-18 1a4b9c7
DOC: tweak to paper
