### 2020-02-18 #1959 @jasmainak
Hi @arokem following the comments of @professoralkmin and @ryEllison, we have made the following improvements:

- reorganized the paper into sections to highlight the main contributions of the software.
- added an example in the paper that shows how Pyglmnet can be used on the community crime dataset.
- updated the example in the README document to provide console output and visualizations to assist the first time user.

We believe these should address the main concerns raised by the reviewers. Thanks again for helping us improving the software and the paper
