### 2020-02-18 #1959 @arokem
@jasmainak : one small question/suggestion. The paper emphasizes compatibility with scikit learn. Are the estimators implemented in the package tested with sklearn's `check_estimator` (see: https://scikit-learn.org/stable/developers/develop.html#rolling-your-own-estimator)? That would guarantee compatibility with other sklearn features. I couldn't find that in the software tests.
