### 2020-02-18 035e613
Minor update README example
