### 2020-02-18 glm-tools/pyglmnet#363 @titipata
Test `GLM` with scikit-learn `check_estimator` (setting attributes during init )
According to https://github.com/openjournals/joss-reviews/issues/1959#issuecomment-587441895, we cannot use scikit-learn's `check_estimator` with `GLM` class. We might have set the initial values [here](https://github.com/glm-tools/pyglmnet/blob/master/pyglmnet/pyglmnet.py#L553-L556) and [here](https://github.com/glm-tools/pyglmnet/blob/master/pyglmnet/pyglmnet.py#L562) somewhere outside the init.

``` py
from pyglmnet import GLM
from sklearn.utils.estimator_checks import check_estimator

check_estimator(GLM)
```

Output:
``` py
AssertionError                            Traceback (most recent call last)
<ipython-input-10-96431abb45f5> in <module>
----> 1 check_estimator(GLM)

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/estimator_checks.py in check_estimator(Estimator)
    292         estimator = Estimator()
    293         check_parameters_default_constructible(name, Estimator)
--> 294         check_no_attributes_set_in_init(name, estimator)
    295     else:
    296         # got an instance

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/estimator_checks.py in check_no_attributes_set_in_init(name, estimator)
   2067             "Estimator %s should not set any attribute apart"
   2068             " from parameters during init. Found attributes %s."
-> 2069             % (name, sorted(invalid_attr)))
   2070     # Ensure that each parameter is set in init
   2071     invalid_attr = set(init_params) - set(vars(estimator)) - {"self"}

AssertionError: Estimator GLM should not set any attribute apart from parameters during init. Found attributes ['beta0_', 'beta_', 'n_iter_', 'rng', 'ynull_'].
```
