### 2020-02-18 glm-tools/pyglmnet#363 @jasmainak
I fear this will go down a rabbit hole, but maybe good for the long term. I'll give it a try later today and see how far I can go ...
