### 2020-02-18 glm-tools/pyglmnet#363 @titipata
We can also fix in the writing (`paper.md`) first where we specify some functionality that can be used with scikit-learn but not all.
