### 2020-02-18 glm-tools/pyglmnet#363 @titipata
@jasmainak I did try moving

``` py
self.beta0_ = None
self.beta_ = None
self.ynull_ = None
self.n_iter_ = 0
self.rng = np.random.RandomState(self.random_state)
```

into `fit(...)`. However, I'm getting a following error from `check_estimator` instead:

``` py
<ipython-input-2-96431abb45f5> in <module>
----> 1 check_estimator(GLM)

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/estimator_checks.py in check_estimator(Estimator)
    300     for check in _yield_all_checks(name, estimator):
    301         try:
--> 302             check(name, estimator)
    303         except SkipTest as exception:
    304             # the only SkipTest thrown currently results from not

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/testing.py in wrapper(*args, **kwargs)
    353             with warnings.catch_warnings():
    354                 warnings.simplefilter("ignore", self.category)
--> 355                 return fn(*args, **kwargs)
    356
    357         return wrapper

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/estimator_checks.py in check_fit_score_takes_y(name, estimator_orig)
   1101         func = getattr(estimator, func_name, None)
   1102         if func is not None:
-> 1103             func(X, y)
   1104             args = [p.name for p in signature(func).parameters.values()]
   1105             if args[0] == "self":

~/Documents/github/pyglmnet/pyglmnet/pyglmnet.py in fit_predict(self, X, y)
    957             The predicted targets of shape (n_samples,).
    958         """
--> 959         return self.fit(X, y).predict(X)
    960
    961     def score(self, X, y):

~/Documents/github/pyglmnet/pyglmnet/pyglmnet.py in fit(self, X, y)
    786
    787         if hasattr(self, '_allow_refit') and self._allow_refit is False:
--> 788             raise ValueError("This glm object has already been fit before."
    789                              "A refit is not allowed")
    790

ValueError: This glm object has already been fit before.A refit is not allowed
```
