### 2020-02-18 a885f13
Attempt to make GLM with check_estimator
