### 2020-02-18 glm-tools/pyglmnet#363 @titipata
@jasmainak so last point is about scipy's `expit`. After I commented out `fit_predict`, I got the following error:

``` py
<ipython-input-5-96431abb45f5> in <module>
----> 1 check_estimator(GLM)

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/estimator_checks.py in check_estimator(Estimator)
    300     for check in _yield_all_checks(name, estimator):
    301         try:
--> 302             check(name, estimator)
    303         except SkipTest as exception:
    304             # the only SkipTest thrown currently results from not

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/estimator_checks.py in check_complex_data(name, estimator_orig)
    671     estimator = clone(estimator_orig)
    672     assert_raises_regex(ValueError, "Complex data not supported",
--> 673                         estimator.fit, X, y)
    674
    675

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/_unittest_backport.py in assertRaisesRegex(self, expected_exception, expected_regex, *args, **kwargs)
    222         context = _AssertRaisesContext(expected_exception,
    223                                        self, expected_regex)
--> 224         return context.handle('assertRaisesRegex', args, kwargs)

~/anaconda3/lib/python3.7/site-packages/sklearn/utils/_unittest_backport.py in handle(self, name, args, kwargs)
    111                 self.obj_name = str(callable_obj)
    112             with self:
--> 113                 callable_obj(*args, **kwargs)
    114         finally:
    115             # bpo-23890: manually break a reference cycle

~/Documents/github/pyglmnet/pyglmnet/pyglmnet.py in fit(self, X, y)
    832                                     alpha, self.Tau,
    833                                     reg_lambda, X, y, self.eta,
--> 834                                     beta, self.fit_intercept)
    835                 # Update
    836                 beta = beta - self.learning_rate * grad

~/Documents/github/pyglmnet/pyglmnet/pyglmnet.py in _grad_L2loss(distr, alpha, Tau, reg_lambda, X, y, eta, beta, fit_intercept)
    239
    240     z = _z(beta[0], beta[1:], X, fit_intercept)
--> 241     mu = _mu(distr, z, eta, fit_intercept)
    242     grad_mu = _grad_mu(distr, z, eta)
    243

~/Documents/github/pyglmnet/pyglmnet/pyglmnet.py in _mu(distr, z, eta, fit_intercept)
    100         mu = z
    101     elif distr == 'binomial':
--> 102         mu = expit(z)
    103     elif distr == 'probit':
    104         mu = norm.cdf(z)

TypeError: ufunc 'expit' not supported for the input types, and the inputs could not be safely coerced to any supported types according to the casting rule ''safe''
```

Not sure how to fix this properly.
