### 2020-02-18 51df734
Adding check_estimator in test script
