### 2020-02-18 42e9bc1
Follow scikit-learn template for estimator
