### 2020-02-18 5d12206
Try using BaseEstimator from scikit-learn
