### 2020-02-18 d1219ae
Add _ to rng according to scikit-learn
