### 2020-02-18 240cf05
Adding back fit_predict method
