### 2020-02-18 12fb8d1
Uncomment _allow_refit
