### 2020-02-18 2d441b6
Comment back _allow_refit
