### 2020-02-18 1c7d244
Comment back _allow_refit
