### 2020-02-18 86f5a1c
Further use check_random_state from scikit-learn
