### 2020-02-18 c47b78f
Further use scikit-learn random state convention
