### 2020-02-18 glm-tools/pyglmnet#363 @titipata
As @jasmainak suggested, using sklearn dependencies make `GLM` works with `check_estimator` (see #364). We can discard the above errors.
