### 2020-02-19 1c54394
Polishing test and pyglmnet class
