### 2020-02-19 0937c85
Remove assignment of is_fitted_ in GLMCV
