### 2020-02-19 2a4dddd
Adapt score function and add test score
