+++ tests/test_pyglmnet.py
@@ -15,6 +15,8 @@ from sklearn.preprocessing import StandardScaler
 from sklearn.model_selection import GridSearchCV, cross_val_score, KFold
 from sklearn.linear_model import ElasticNet
 from sklearn.utils.estimator_checks import check_estimator
+from sklearn.utils.testing import assert_raises
+from sklearn.utils.validation import NotFittedError
 
 from pyglmnet import (GLM, GLMCV, _grad_L2loss, _L2loss, simulate_glm,
                       _gradhess_logloss_1d, _loss, datasets, ALLOWED_DISTRS)
@@ -497,6 +499,22 @@ def test_api_input():
     with pytest.warns(UserWarning, match='Reached max number of iterat'):
         glm.fit(X, y)
 
+
+def test_score_before_fitting():
+    """Test thaat score function cannot be applied before fitting the model."""
+
+    random_state = 1
+    state = np.random.RandomState(random_state)
+    n_samples, n_features = 100, 5
+
+    X = state.normal(0, 1, (n_samples, n_features))
+    y = state.normal(0, 1, (n_samples, ))
+
+    glm = GLM(distr='gaussian')
+    
+    with assert_raises(NotFittedError, match="Cannot apply score before fitting"):
+        glm.score(X, y)
+    
 
 def test_intro_example():
     """Test that the intro example works."""
