### 2020-02-19 5f3b3b2
Somehow default attribute should be binomial/probit for check_estimator to work
