### 2020-02-19 81ee655
Update documentation according to JOSS paper
