+++ doc/start.rst
@@ -39,3 +39,7 @@ Here is a brief example of how to use the ``GLM()`` class.
 
     # score the model on test data
     deviance = glm.score(Xtest, ytest)
+
+``GLM()`` class is also designed to be compatible with ``scikit-learn``. 
+This means that you can apply cross-validation or performing grid search over 
+hyperparameters using scikit-learn's ``KFold`` or ``GridSearchCV``.
