### 2020-02-19 09da399
Update documentation according to JOSS paper
