+++ pyglmnet/externals/sklearn/__init__.py
@@ -1,0 +1,33 @@
+"""
+Machine learning module for Python
+==================================
+sklearn is a Python module integrating classical machine
+learning algorithms in the tightly-knit world of scientific Python
+packages (numpy, scipy, matplotlib).
+It aims to provide simple and efficient solutions to learning problems
+that are accessible to everybody and reusable in various contexts:
+machine-learning as a versatile tool for science and engineering.
+See http://scikit-learn.org for complete documentation.
+"""
+
+from . import utils
+
+__all__ = ['utils']
+
+# PEP0440 compatible formatted version, see:
+# https://www.python.org/dev/peps/pep-0440/
+#
+# Generic release markers:
+#   X.Y
+#   X.Y.Z   # For bugfix releases
+#
+# Admissible pre-release markers:
+#   X.YaN   # Alpha release
+#   X.YbN   # Beta release
+#   X.YrcN  # Release Candidate
+#   X.Y     # Final release
+#
+# Dev branch marker is: 'X.Y.dev' or 'X.Y.devN' where N is an integer.
+# 'X.Y.dev0' is the canonical version of 'X.Y.dev'
+#
+__version__ = '0.23.dev0'
