+++ pyglmnet/externals/__init__.py
@@ -1,0 +1,3 @@
+"""
+External, bundled dependencies.
+"""
