+++ pyglmnet/externals/sklearn/utils/config.py
@@ -1,0 +1,24 @@
+"""Global configuration state and functions for management
+"""
+import os
+
+
+_global_config = {
+    'assume_finite': bool(os.environ.get('SKLEARN_ASSUME_FINITE', False)),
+    'working_memory': int(os.environ.get('SKLEARN_WORKING_MEMORY', 1024)),
+    'print_changed_only': False,
+}
+
+
+def get_config():
+    """Retrieve current values for configuration set by :func:`set_config`
+    Returns
+    -------
+    config : dict
+        Keys are parameter names that can be passed to :func:`set_config`.
+    See Also
+    --------
+    config_context: Context manager for global scikit-learn configuration
+    set_config: Set global scikit-learn configuration
+    """
+    return _global_config.copy()
