### 2020-02-20 c71a279
Minor externals folder fix
