### 2020-02-20 1ee1cdb
Flake8 corrections
