+++ pyglmnet/externals/sklearn/exceptions.py
@@ -4,6 +4,7 @@ classes used across scikit-learn.
 """
 
 __all__ = ['NotFittedError',
+           'DataConversionWarning',
            'NonBLASDotWarning']
 
 
@@ -21,6 +22,21 @@ class NotFittedError(ValueError, AttributeError):
     ...     print(repr(e))
     NotFittedError("This LinearSVC instance is not fitted yet. Call 'fit' with
     appropriate arguments before using this estimator."...)
+    .. versionchanged:: 0.18
+       Moved from sklearn.utils.validation.
+    """
+
+
+class DataConversionWarning(UserWarning):
+    """Warning used to notify implicit data conversions happening in the code.
+    This warning occurs when some input data needs to be converted or
+    interpreted in a way that may not match the user's expectations.
+    For example, this warning may occur when the user
+        - passes an integer array to a function which expects float input and
+          will convert the input
+        - requests a non-copying operation, but a copy is required to meet the
+          implementation's data-type expectations;
+        - passes an input whose shape can be interpreted ambiguously.
     .. versionchanged:: 0.18
        Moved from sklearn.utils.validation.
     """
