### 2020-02-20 a017a68
Flake8 corrections
