### 2020-02-20 986308e
Removing unused lines
