### 2020-02-20 cbaa4a6
Adding whats new for sklearn externals
