+++ doc/whats_new.rst
@@ -56,6 +56,7 @@ BUG
 API
 ~~~
 
+    - Refactored `GLM` class so that it is fully support by `scikit-learn` by `Titipat Achakulvisut`_.
     - Refactored `GLM` class into two classes: :class:`pyglmnet.GLM` for fitting a single regularization parameter,
       and :class:`pyglmnet.GLMCV` for fitting along a regularization path with cross-validation and selection of best
       regularization parameter by `Pavan Ramkumar`_ and `Mainak Jas`_.
