### 2020-02-20 286abea
Change back to default param, remove a line in config, update whats new
