+++ pyglmnet/externals/sklearn/utils/config.py
@@ -2,7 +2,6 @@
 """
 import os
 
-
 _global_config = {
     'assume_finite': bool(os.environ.get('SKLEARN_ASSUME_FINITE', False)),
     'working_memory': int(os.environ.get('SKLEARN_WORKING_MEMORY', 1024)),
