### 2020-02-20 4dd6708
Fix typos in start documentation
