### 2020-02-20 24fa8fd
Work around for predict_proba, allows allow but return predict for the other cases
