### 2020-02-20 38e095b
use UserWarning instead, remove unused test line
