### 2020-02-20 2bec2e3
use UserWarning instead, remove unused test line
