+++ doc/whats_new.rst
@@ -19,10 +19,11 @@ BUG
 
     - Graceful handling of small Hessian term in coordinate descent solver that led to exploding update term by `Pavan Ramkumar`_.
 
+    - Ensure full compatibility of `GLM` class with `scikit-learn` by `Titipat Achakulvisut`_.
+
 API
 ~~~
 
-    - Refactored `GLM` class so that it is fully support by `scikit-learn` by `Titipat Achakulvisut`_.
 
 .. _1_0_1:
 
