### 2020-02-20 8d2230a
Update start to be the same as in README.rst
