### 2020-02-20 #1959 @jasmainak
They weren't but now we added this test thanks to @titipata in https://github.com/glm-tools/pyglmnet/pull/364. We were indirectly testing this in many ways but I agree `check_estimator` is the best way to guarantee compatibility. Thanks a lot for the suggestion!
