### 2020-02-20 #1959 @arokem
Awesome! I think that the paper is ready for a final review by the EICs. Before we go ahead, could you please:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
