### 2020-02-25 #1959 @jasmainak
@arokem here you go!

- the version tag of the [archived release on github](https://github.com/glm-tools/pyglmnet/releases/tag/1.1.1) is 1.1.1
- doi of the archived version on zenodo is: 10.5281/zenodo.3686564
- the author list and title of the deposit matches the author list and title on the paper
