### 2020-02-25 c3ef2ca
Add Pillow's tutorial for RGC spikes prediction
