### 2020-02-25 9bb6976
Add Pillow's tutorial for RGC spikes prediction
