### 2020-02-25 34c24de
Update text in the tutorial
