### 2020-02-25 7350072
Prediction with spike history
