### 2020-02-25 e37e927
Poisson GLM with spikes couple
