### 2020-02-25 7bb016c
Minor fix on spikes couple
