### 2020-02-25 81f9934
Catch on the design matrix include its prediction
