### 2020-02-26 53b4b15
Trim examples, adding license when downloading data
