### 2020-02-26 a4133df
Cleaning text in the tutorial
