+++ examples/plot_glm_retinal_ganglia.py
@@ -6,19 +6,21 @@ GLM for Spike Trains Prediction in Primate Retinal Ganglion Cells
 
 * Original tutorial adapted from Johnathan Pillow, Princeton University
 * Dataset provided by E.J. Chichilnisky, Stanford University
-* **The dataset is granted by the original authors for the educational use in this tutorial**
-* If you want to use it beyond educational purposes, please contact ``pillow@princeton.edu``
+* The dataset is granted by the original authors for only educational use
+* Please contact ``pillow@princeton.edu`` if using beyond its purposes
 
-First of all, we would like to thank Professor Johnathan Pillow and Professor E.J. Chichilnisky
-for sharing their tutorial dataset. The original MATLAB and Python tutorial can be found from
+First of all, we would like to thank Professor Johnathan Pillow
+and Professor E.J. Chichilnisky for sharing their tutorial dataset.
+The original MATLAB and Python tutorial can be found from
 https://github.com/pillowlab/GLMspiketraintutorial.
 
 These data were collected by Valerie Uzzell in the lab of
 E.J. Chichilnisky at the Salk Institute.  For full information see
 Uzzell et al. (J Neurophys 04) [1]_, or Pillow et al. (J Neurosci 2005) [2]_.
 
-In this tutorial, we will demonstrate how to fit linear GLM and linear-nonlinear GLM 
-(a.k.a Poisson GLM) to predict the spike counts recorded from primate retinal ganglion cells.
+In this tutorial, we will demonstrate how to fit linear GLM and
+linear-nonlinear GLM (a.k.a Poisson GLM) to predict the spike counts
+recorded from primate retinal ganglion cells.
 The dataset contains spike responses from 2 ON and 2 OFF parasol
 retinal ganglion cells (RGCs) in primate retina, stimulated with
 full-field `binary white noise`. Two experiment performed consisted of a
@@ -28,10 +30,12 @@ which can be used for computing the spike-triggered average
 
 References
 ----------
-.. [1] Uzzell, V. J., and E. J. Chichilnisky. "Precision of spike trains in primate
-   retinal ganglion cells." Journal of Neurophysiology 92.2 (2004)
-.. [2] Pillow, Jonathan W., et al. "Prediction and decoding of retinal ganglion cell
-   responses with a probabilistic spiking model." Journal of Neuroscience 25.47 (2005)
+.. [1] Uzzell, V. J., and E. J. Chichilnisky. "Precision of
+   spike trains in primate retinal ganglion cells."
+   Journal of Neurophysiology 92.2 (2004)
+.. [2] Pillow, Jonathan W., et al. "Prediction and decoding of
+   retinal ganglion cell responses with a probabilistic
+   spiking model." Journal of Neuroscience 25.47 (2005)
 
 """
 
@@ -46,7 +50,7 @@ import os.path as op
 import json
 
 import numpy as np
-from numpy.linalg import inv, norm
+from numpy.linalg import inv
 from scipy.linalg import hankel
 
 from pyglmnet import GLM
@@ -56,7 +60,7 @@ from tempfile import TemporaryDirectory
 
 ########################################################
 #
-# Fetch the dataset. The JSON file contains the 
+# Fetch the dataset. The JSON file contains the
 # following keys:  ``stim`` (binary stochastic stimulation),
[…]
