### 2020-02-26 #1959 @arokem
@openjournals/joss-eics : I believe that this is ready for your review
