### 2020-02-27 155393b
Adding parameter documentation for RGC dataset
