+++ pyglmnet/__init__.py
@@ -18,6 +18,7 @@
 
 __version__ = '1.2.dev0'
 
+
 from .pyglmnet import GLM, GLMCV, _grad_L2loss, _L2loss, simulate_glm, _gradhess_logloss_1d, _loss, ALLOWED_DISTRS
 from .utils import softmax, label_binarizer, set_log_level
 from .datasets import fetch_tikhonov_data, fetch_RGCs_data
