### 2020-02-27 c01ae3a
Adding parameter documentation for RGC dataset
