### 2020-02-27 132f578
Resolve authorship
