### 2020-02-27 4198f9c
Remove STA, fix text according to comments
