+++ examples/plot_glm_retinal_ganglia.py
@@ -17,7 +17,7 @@ E.J. Chichilnisky at the Salk Institute. For full information see
 Uzzell et al. (J Neurophys 04) [1]_, or Pillow et al. (J Neurosci 2005) [2]_.
 
 In this tutorial, we will demonstrate how to fit linear GLM and
-linear-nonlinear GLM (a.k.a Poisson GLM) to predict the spike counts
+Poisson GLM to predict the spike counts
 recorded from primate retinal ganglion cells.
 The dataset contains spike responses from 2 ON and 2 OFF parasol
 retinal ganglion cells (RGCs) in primate retina, stimulated with
@@ -84,7 +84,7 @@ spike_times = [
 n_cells = len(spike_times)  # total number of cells
 dt = stim_times[1] - stim_times[0]  # time between the stimulation
 n_t = len(stim)  # total number of the stimulation
-f = 1. / dt  # frequency of the stimulation
+sfreq = 1. / dt  # frequency of the stimulation
 
 ########################################################
 #
@@ -94,6 +94,10 @@ f = 1. / dt  # frequency of the stimulation
 cell_idx = 2  # pick cell number 2 (ON cell)
 spike_time = spike_times[cell_idx]  # pick spike time to work with
 n_spikes = len(spike_time)  # number of spikes
+
+# bin the spikes
+t_bins = np.arange(n_t + 1) * dt
+spikes_binned, _ = np.histogram(spike_time, t_bins)
 
 print('Loaded RGC data: cell {}'.format(cell_idx))
 print('Number of stim frames: {:d} ({:.1f} minutes)'.
@@ -106,26 +110,34 @@ print('Number of spikes: {} (mean rate = {:.1f} Hz)\n'.
 sample_index = np.arange(120)
 t_sample = dt * sample_index
 
-plt.subplot(2, 1, 1)
+plt.subplot(3, 1, 1)
 plt.step(t_sample, stim[sample_index])
-plt.title('Raw Stimulus (full field flicker)'
-          ' and corresponding spike times')
+plt.xlim([t_sample.min(), t_sample.max()])
+plt.title('Raw Stimulus, spikes, and'
+          ' spike counts')
 plt.ylabel('Stimulation Intensity')
 
-plt.subplot(2, 1, 2)
+plt.subplot(3, 1, 2)
 tspplot = spike_time[(spike_time >= t_sample.min()) &
                      (spike_time < t_sample.max())]
 plt.stem(spike_time[sample_index],
          [1] * len(spike_time[sample_index]))
 plt.xlim([t_sample.min(), t_sample.max()])
 plt.ylim([0, 2])
+plt.ylabel('Spikes')
+
+plt.subplot(3, 1, 3)
+markerline, _, _ = plt.stem(t_sample, spikes_binned[sample_index])
+markerline.set_markerfacecolor('none')
+plt.xlim([t_sample.min(), t_sample.max()])
 plt.xlabel('Time (s)')
-plt.ylabel('Spikes (0 or 1)')
+plt.ylabel('Binned Spike counts')
 plt.show()
 
 ########################################################
 #
 # We can use ``scipy``'s function ``hankel`` to make our design matrix.
+# Design matrix can be created using the stimulation.
 
 n_t_filt = 25  # tweak this to see different results
 stim_padded = np.pad(stim, (n_t_filt - 1, 0))
@@ -139,31 +151,6 @@ plt.xlabel('lags before spike time')
 plt.ylabel('time bin of response')
 plt.title('Sample first 50 rows of design'
           ' matrix created using Hankel')
-plt.colorbar()
-plt.show()
-
-########################################################
[…]
