### 2020-02-27 dd84455
Adding text about how to create design matrix
