### 2020-02-28 cbd6a2a
Fix multiple-line equations
