+++ doc/tutorial.rst
@@ -167,12 +167,10 @@ To calculate the gradients of the cost function with respect to :math:`\beta_0` 
 
 .. math::
 
-     \begin{eqnarray}
-         J(\beta_0, \beta) &= \sum_i \bigg\{ \log(1 + \exp(\beta_0 + \beta^T x_i))\\
-           & - y_i \log(\log(1 + \exp(\beta_0 + \beta^T x_i)))\bigg\}\\
-           & + \lambda(1 - \alpha)\frac{1}{2} \|\beta\|^2_{\mathcal{l_2}}
-           + \lambda\alpha\|\beta\|_{\mathcal{l_1}}
-     \end{eqnarray}
+    J(\beta_0, \beta) &= \sum_i \bigg\{ \log(1 + \exp(\beta_0 + \beta^T x_i))\\
+    &- y_i \log(\log(1 + \exp(\beta_0 + \beta^T x_i)))\bigg\}\\
+    &+ \lambda(1 - \alpha)\frac{1}{2} \|\beta\|^2_{\mathcal{l_2}}
+    + \lambda\alpha\|\beta\|_{\mathcal{l_1}}
 
 
 Since we will apply coordinate descent, let's rewrite this cost in terms of each
@@ -180,11 +178,9 @@ scalar parameter :math:`\beta_j`
 
 .. math::
 
-     \begin{eqnarray}
-         J(\beta_0, \beta) &= \sum_i \bigg\{ \log(1 + \exp(\beta_0 + \sum_j \beta_j x_{ij}))
-         & - y_i \log(\log(1 + \exp(\beta_0 + \sum_j \beta_j x_{ij})))\bigg\}\\
-         & + \lambda(1-\alpha)\frac{1}{2} \sum_j \beta_j^2 + \lambda\alpha\sum_j \mid\beta_j\mid
-     \end{eqnarray}
+    J(\beta_0, \beta) &= \sum_i \bigg\{ \log(1 + \exp(\beta_0 + \sum_j \beta_j x_{ij}))
+    &- y_i \log(\log(1 + \exp(\beta_0 + \sum_j \beta_j x_{ij})))\bigg\}\\
+    &+ \lambda(1-\alpha)\frac{1}{2} \sum_j \beta_j^2 + \lambda\alpha\sum_j \mid\beta_j\mid
 
 Let's take the derivatives of some big expressions using chain rule.
 Define :math:`z_i = \beta_0 + \sum_j \beta_j x_{ij}`.
@@ -193,19 +189,17 @@ For the nonlinearity in the first term :math:`y = q(z) = \log(1+e^{z(\theta)})`,
 
 .. math::
 
-     \begin{eqnarray}
-     \frac{\partial y}{\partial \theta} &= \frac{\partial q}{\partial z}\frac{\partial z}{\partial \theta}\\
-     & = \frac{e^z}{1+e^z}\frac{\partial z}{\partial \theta}\\
-     & = \sigma(z)\frac{\partial z}{\partial \theta}
-     \end{eqnarray}
+    \frac{\partial y}{\partial \theta} &= \frac{\partial q}{\partial z}\frac{\partial z}{\partial \theta}\\
+    &= \frac{e^z}{1+e^z}\frac{\partial z}{\partial \theta}\\
+    &= \sigma(z)\frac{\partial z}{\partial \theta}
 
 For the nonlinearity in the second term :math:`y = \log(q(z)) = \log(\log(1+e^{z(\theta)}))`,
 
 .. math::
 
[…]
