+++ examples/plot_glm_retinal_ganglia.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 """
-=================================================================
-GLM for Spike Trains Prediction in Primate Retinal Ganglion Cells
-=================================================================
+================================================================
+GLM for Spike Train Prediction in Primate Retinal Ganglion Cells
+================================================================
 
 * Original tutorial adapted from Johnathan Pillow, Princeton University
 * Dataset provided by E.J. Chichilnisky, Stanford University
@@ -14,7 +14,7 @@ https://github.com/pillowlab/GLMspiketraintutorial.
 
 These data were collected by Valerie Uzzell in the lab of
 E.J. Chichilnisky at the Salk Institute. For full information see
-Uzzell et al. (J Neurophys 04) [1]_, or Pillow et al. (J Neurosci 2005) [2]_.
+Uzzell et al. [1]_, or Pillow et al. [2]_.
 
 In this tutorial, we will demonstrate how to fit linear GLM and
 Poisson GLM to predict the spike counts
@@ -73,15 +73,7 @@ with TemporaryDirectory(prefix="tmp_glm-tools") as temp_dir:
 stim = np.array(rgcs_dataset['stim'])
 stim_times = np.array(rgcs_dataset['stim_times'])
 
-# spike times for all 4 cells (0-1 are OFF cells, 2-3 are ON cells)
-spike_times = [
-    np.array(rgcs_dataset['spike_times']['cell_0']),
-    np.array(rgcs_dataset['spike_times']['cell_1']),
-    np.array(rgcs_dataset['spike_times']['cell_2']),
-    np.array(rgcs_dataset['spike_times']['cell_3']),
-]
-
-n_cells = len(spike_times)  # total number of cells
+n_cells = len(rgcs_dataset['spike_times'])
 dt = stim_times[1] - stim_times[0]  # time between the stimulation
 n_t = len(stim)  # total number of the stimulation
 sfreq = 1. / dt  # frequency of the stimulation
@@ -92,7 +84,8 @@ sfreq = 1. / dt  # frequency of the stimulation
 # In this case, we will pick cell number 2 (ON cell).
 
 cell_idx = 2  # pick cell number 2 (ON cell)
-spike_time = spike_times[cell_idx]  # pick spike time to work with
+spike_time = rgcs_dataset['spike_times']['cell_%d' % cell_idx]
+spike_time = np.array(spike_time)
 n_spikes = len(spike_time)  # number of spikes
 
 # bin the spikes
@@ -106,32 +99,34 @@ print('Time bin size: {:.1f} ms'.format(dt * 1000))
 print('Number of spikes: {} (mean rate = {:.1f} Hz)\n'.
       format(n_spikes, n_spikes / n_t * 60))
 
-# sample indices for visualization
-sample_index = np.arange(120)
-t_sample = dt * sample_index
+########################################################
+#
+# Visualize first 120 samples
+sample_idx = np.arange(120)
+t_sample = dt * sample_idx
+tmax = t_sample.max()
 
-plt.subplot(3, 1, 1)
-plt.step(t_sample, stim[sample_index])
-plt.xlim([t_sample.min(), t_sample.max()])
-plt.title('Raw Stimulus, spikes, and'
-          ' spike counts')
-plt.ylabel('Stimulation Intensity')
+fig, axes = plt.subplots(3, 1, sharex=True)
+axes[0].step(t_sample, stim[sample_idx])
+axes[0].set_xlim([0., tmax])
+axes[0].set_title('Raw Stimulus, spikes, and'
+                  ' spike counts')
+axes[0].set_ylabel('Stimulation Intensity')
 
[…]
