+++ .circleci/config.yml
@@ -1,6 +1,6 @@
 version: 2
 jobs:
-    build:
+    build_docs:
         docker:
             - image: circleci/python:3.7.0-stretch-node-browsers
         steps:
@@ -36,3 +36,9 @@ jobs:
             - store_artifacts:
                 path: doc/_build/html/
                 destination: html
+
+workflows:
+  version: 2
+  search_build:
+    jobs:
+      - build_docs
