### 2020-02-28 1b9b3ce
Change function name, update variables name
