### 2020-02-29 ed1cc8c
align multiple-line equations properly
