### 2020-02-29 19d3a4e
fix equation alignment
