### 2020-02-29 6e21609
align multiple-line equations properly (#372)
* align multiple-line equations properly

* fix equation alignment
